requirements.txt
Django==1.10.7
django-registration-redux==1.3
